made component SelectFilter, moved filtercode to component
