+++ src/components/ProductsListContainer.js
@@ -4,6 +4,7 @@ import { getProducts, cartAdd, filterSearch } from "../actions/productActions";
 import { Pacman } from "react-pure-loaders";
 
 import ProductBox from "./ProductBox";
+import SelectFilter from "./SelectFilter";
 
 export default function ProductListContainer() {
   const [categoryFilter, setCategoryFilter] = useState("all");
@@ -22,17 +23,6 @@ export default function ProductListContainer() {
     dispatch(cartAdd(id));
   };
 
-  const handleFilter = (event) => {
-    if (event.target.value === "all") {
-      dispatch(getProducts());
-      setCategoryFilter("all");
-      setCurrentPage(1);
-    } else {
-      dispatch(filterSearch(event.target.value));
-      setCategoryFilter(event.target.value);
-      setCurrentPage(1);
-    }
-  };
   const firstPage = () => {
     if (currentPage > 1) {
       setCurrentPage(1);
@@ -71,24 +61,12 @@ export default function ProductListContainer() {
   const totalPages = products && Math.ceil(productsList.length / productsPerPage);
 
   return (
-    <section className="top-product">
+    <section className="">
       <div className="pagina-naam">
         <h2>Producten</h2>
       </div>
-      <span className="filter">
-        <label>Filter producten:</label>
-
-        <select id="category" onChange={handleFilter}>
-          <option value="all">Alle Producten</option>
-          <option value={"Bijtringen & Rammelaars"}>
-            {"Bijtringen & Rammelaars"}
-          </option>
-          <option value="Baby Overige">Baby Overige</option>
-          <option value="Knuffels">Knuffels</option>
-          <option value="Sleutelhangers">Sleutelhangers</option>
-          <option value="Overige">Overige</option>
-        </select>
-      </span>
+    
+      <SelectFilter setCategoryFilter={setCategoryFilter} dispatch={dispatch} setCurrentPage={setCurrentPage} />
 
       <div>
         {currentProducts.length === 0 ? (
