+++ src/style/global.scss
@@ -300,6 +300,7 @@ img {
 }
 
 .pagina-naam {
+  margin-top: 25px;
   background: $wit;
   text-align: center;
 }
